PANDO-21: Move `IPandoSerializer` to `Pando.Serialization` namespace
