+++ docs/PandoExampleProject/ChessTests.cs
@@ -2,6 +2,7 @@ using System.Collections.Immutable;
 using System.Text.RegularExpressions;
 using Pando;
 using Pando.DataSources;
+using Pando.Serialization;
 using PandoExampleProject.Serializers;
 using Xunit;
