+++ docs/PandoExampleProject/Serializers/ChessPieceSerializer.cs
@@ -1,6 +1,6 @@
 using System;
-using Pando;
 using Pando.DataSources;
+using Pando.Serialization;
 
 namespace PandoExampleProject.Serializers;
