+++ src/Pando/DataSources/PersistenceBackedDataSource.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Immutable;
 using Pando.DataSources.Utils;
+using Pando.Serialization;
 
 namespace Pando.DataSources;
