+++ tests/PandoTests/Utils/ToArrayReader.cs
@@ -1,7 +1,7 @@
 using System;
-using Pando;
 using Pando.DataSources;
 using Pando.DataSources.Utils;
+using Pando.Serialization;
 
 namespace PandoTests.Utils;
